change mobile categorizations to toggle buttons, update with of mobile contents
